+++ src/app/property/[id]/detail/InfoBox.tsx
