+++ src/app/property/[id]/detail/SectionList.tsx
