+++ src/app/property/[id]/detail/components/GrayButton.tsx
