+++ src/app/property/[id]/detail/components/ScrollableTabBar.tsx
