+++ src/app/property/[id]/detail/section/DealSection.tsx
