+++ src/app/property/[id]/detail/section/InfoSection.tsx
