+++ src/app/property/[id]/detail/section/LocationSection.tsx
