+++ src/app/property/[id]/detail/section/PriceSection.tsx
