+++ src/app/property/[id]/review/EnvironmentTabs.tsx
