+++ src/app/property/[id]/review/[reviewId]/page.tsx
@@ -1,0 +1,1 @@
+"use client";
